Fix for Issue #49
GtfReader now skipps annotations in the annotation file that don't
relate to the chromosome that is currently being viewed.

This check assumes that the chromosome name at the beginning of the
annotation file matches to the substring of the chromosome file name
from "chr" to the end minus the file extension.

TODO: Make a UI variable that the user can define a chromosome name to
read from the annotation.