Fasta Reader Update Round 1
Rewrite of the Fasta Reader to improve memory performance.
This is round one, so it it still has some refactoring to go through.
More parts will be pulled out into methods and there is code that can be
shrunk and optimized.

The progress bar currently does not function as intended.