Screen Capture Framework
As per Issue #61 created initial framework for taking screen shots.
The current method is to go to the graphics card buffer and read in
pixel by pixel and convert to a QImage.

This code should be refactored out into methods.

TODO: Save under variable file name and add into user variables as an
overlay to the image.